Extract CardContent component from MovieCard component.
